search bar fixed common in every component,Print button settled in common component
