+++ src/features/accountant/Bank/BankHome.tsx
@@ -1,12 +1,14 @@
 import Ellipsis from "../../../assets/icons/Ellipsis"
 import NewBankModal from "./NewBankModal"
-import SearchBar from "../../sales/SearchBar"
+import SearchBar from "../../../Components/SearchBar"
 import Table from "./Table"
+import { useState } from "react"
 
 
 type Props = {}
 
 function BankHome({}: Props) {
+  const [searchValue, setSearchValue] = useState<string>("");
   return (
     <>
     <div className="p-5 bg-slate-50 h-[100vh]">
@@ -29,7 +31,7 @@ function BankHome({}: Props) {
 
     <div className="mt-5 bg-white">
     <div className="p-6">
-        <SearchBar placeholder="Search Accounts"/>
+        <SearchBar  onSearchChange={setSearchValue} searchValue={searchValue} placeholder="Search Accounts"/>
     </div>
     <div>
         <Table/>
