+++ src/features/inventory/Item/ItemHome.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import PlusCircle from "../../../assets/icons/PlusCircle";
 import Button from "../../../Components/Button";
-import Print from "../../sales/Print";
+import Print from "../../../Components/PrintButton";
 import SearchBar from "../../../Components/SearchBar";
 import Items from "./Items";
 import ItemSort from "./ItemSort";
@@ -9,7 +9,7 @@ import ItemTable from "./ItemTable";
 type Props = {}
 
 function ItemHome({}: Props) {
-    const [searchValue, setSearchValue] = useState<string>("");
+  const [searchValue, setSearchValue] = useState<string>("");
   return (
     <div className="m-7">
     <div className="flex justify-between items-center">
