+++ src/features/purchase/debitNote/DebitNote.tsx
@@ -1,16 +1,18 @@
 import { Link } from "react-router-dom"
 import PlusCircle from "../../../assets/icons/PlusCircle"
 import Button from "../../../Components/Button"
-import SearchBar from "../../sales/SearchBar"
+import SearchBar from "../../../Components/SearchBar"
 import DebitDropdown from "./DebitDropdown"
 import DebitPrint from "./DebitPrint"
 import DebitSort from "./DebitSort"
 import DebitTable from "./DebitTable"
 import TypesOfDebits from "./TypesOfDebits"
+import { useState } from "react"
 
 type Props = {}
 
 function DebitNote({}: Props) {
+  const [searchValue, setSearchValue] = useState<string>("");
   return (
     <>
          <div className=" px-6 flex items-center relative">
@@ -38,7 +40,7 @@ function DebitNote({}: Props) {
           </div>
           <div className="flex pl-3 pr-3 items-center gap-4 mt-2">
             <div className="w-[80%]">
-              <SearchBar placeholder="Search Sales Order" />
+              <SearchBar  onSearchChange={setSearchValue} searchValue={searchValue} placeholder="Search Sales Order" />
             </div>
             <DebitSort/>
             <DebitPrint/>
