+++ src/features/sales/dashboard/SalesDashboard.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+function SalesDashboard({}: Props) {
+  return (
+    <div>SalesDashboard</div>
+  )
+}
+
+export default SalesDashboard
